[REFACTOR] Move TIR op kind analysis to relax as it is relax oriented
This also keep TIR mostly independent from higher-level IR.
